chore: Publish `near-vm-runner` and it's dependencies (#8782)
My previous PR #8762 was reverted by #8780 because release automation was broken.

Looks like the cause of broken automation was a crate `near-test-contract`. We had a discussion about this package and the reasons to publish it in the first place in the very first PR. Now we have proof that a dependency that mentioned in the `dev-dependencies` section of `Cargo.toml` is not required to be published to crates.io.

Once figured out I created a PR #8779 but it was unnoticed and instead all the changes were reverted in #8780.

As agreed with @jakmeier I am creating the same changes but disabled `near-test-contract` for publishing necessary crates in the future.