Update clap to 4-series (and async-recursion, async-stream, serde, openssl-macros, syn) (#8974)
This is something we need to upgrade the toolchain. Largely the update was uneventful, but it also pulls in other new crates, such as syn 2.0, which leads to many duplicated crates while the ecosystem is moving over to the new version.

One observation is that with the new API it is somewhat easy to get things wrong when using the manual clap API, rather than the derive version. Now that it stores the exact type it parsed, you can no longer rely on clap converting stuff for you. You are kind-of required to get types right for yourself, and any mismatch will result in a runtime panic, which is iffy.