fix(state-sync): Always use flat storage when catching up (#9311)
The original code made the use of flat storage conditional on the node tracking that shard this epoch.
If a node prepares to track shard S next epoch E, then it downloads its state (E-1) and applies chunks in order. To apply chunks correctly in a way compatible with the rest of the network, it needs to be using flat storage.

Also add a metric for the latest block processed during catchup.
Also fix `view-state apply-range` tool not to fail because of getting delayed indices.
Also reduce verbosity of the inlining migration.