[resharding] Remove progress counters from resharding as they are not used. (#9334)
Current resharding logic has a set number of parts and a struct to track how many shards have completed "migration". This unfortunately isn't used anywhere.

The struct for storing this data is created in `sync_shards_state_split_scheduling_status`, but I don't see it being used anywhere other than for perhaps logging.

With flat storage resharding implementation, we are not going to have a fixed number of parts and we would have to think of an alternative struct to track progress. For now it makes sense to remove it.