fix(state-sync): Remove `state_fetch_horizon` because it causes nodes to crash (#9380)
The main issue is that `find_sync_hash()` is not consistent with `check_state_needed()`.
`check_state_needed` compares epoch_id of header_head and head.
`find_sync_hash()` moves back a number of blocks before determining which epoch to sync.
It can lead to a node deciding to state sync, and downloading the same state again.

I've observed that it leads to a crash in this complicated scenario:
* State sync to epoch X
* Finalize the state sync and execute block at height H
* Check orphans and unblock a block at height H+1
* At the same time check if a node needs a state sync - determine that yes, and reset Flat Storage to prepare it to execute a block at height H.
* The orhpaned block gets executed, but flat storage doesn't support height H+1, leading to a panic.