fix(mirror): add an extra key to accounts with no keys created by con… (#9445)
…tracts

when we send CreateAccount txs that come from txs where there is no full access key added, we add an extra access key so that we can control the account in the mirrored chain. But we forget to do this for accounts created by contracts, and many accounts on mainnet are created by calling the `near` contract. This is not a huge problem, but would be if we're mirroring traffic where we happen to run into a new validator being created via contract, and then staked, in which case we wouldn't be able to unstake any of their stake actions without this PR